% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_pGRR.R
\name{AM.pGRR}
\alias{AM.pGRR}
\title{Additive model via partial generalized ridge regression}
\usage{
AM.pGRR(y, X, k = ncol(X), q = k)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{k}{the number of explanatory variables}

\item{q}{a vector or an integer which is the number of knots;
if vector, the number of knots is different for each variable}
}
\value{
estimation results
}
\description{
\code{AM.pGRR} This function estimates additive model via partial generalized ridge regression
}
\examples{
#AM.pGRR(y, X)
}
