% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRRMSC.R
\name{GRR.MSC}
\alias{GRR.MSC}
\title{Generalized ridge regression with the optimal ridge parameters based on
model selection criterion minimization method}
\usage{
GRR.MSC(
  y,
  X,
  MSC = c("GCp", "EGCV"),
  alpha = log(n),
  n = length(y),
  tol = 1e-12
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{MSC}{a model selection criterion; GCp or EGCV}

\item{alpha}{a value (>=2) expressing penalty strength for MSC}

\item{n}{sample size}

\item{tol}{tolerance for rank deficient}
}
\value{
estimation results
}
\description{
\code{GRR.MSC} This function provides generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#GRR.MSC(y, X)
}
