% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{GRR.pred}
\alias{GRR.pred}
\title{Prediction}
\usage{
GRR.pred(newX, res)
}
\arguments{
\item{newX}{a matrix of explanatory variables for future observation}

\item{res}{the output of \code{GRR.MSC}}
}
\value{
predictive values
}
\description{
\code{GRR.pred} This function provides predictive values for future observation
}
\examples{
#GRR.pred(newX, res)
}
