% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGRRMSC.R
\name{pGRR.MSC}
\alias{pGRR.MSC}
\title{Partial generalized ridge regression optimized based on model selection criterion minimization method}
\usage{
pGRR.MSC(
  y,
  X,
  Z,
  MSC = c("EGCV", "GCp", "GCV", "Cp", "MCp"),
  alpha = log(n),
  n = length(y),
  tol = 1e-12
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of penalized explanatory variables without intercept}

\item{Z}{a matrix of non-penalized explanatory variables with intercept if need}

\item{MSC}{a model selection criterion; \code{"EGCV"}, \code{"GCp"}, \code{"GCV"}, \code{"Cp"}, or \code{"MCp"}}

\item{alpha}{a value (>=2) expressing penalty strength for \code{MSC} (only when \code{MSC} is \code{"EGCV"} or \code{"GCp"})}

\item{n}{sample size}

\item{tol}{tolerance for rank deficient}
}
\value{
a list object with "pGRR.MSC" class which has the following elements:
\tabular{ll}{
\code{beta} \tab
estimates for \code{X} \cr \tab \cr
\code{gamma} \tab
estimates for \code{Z} \cr \tab \cr
\code{fitted.values} \tab
fitted values \cr \tab \cr
\code{R2} \tab
the coefficient of determination \cr
}
}
\description{
\code{pGRR.MSC} This function provides partial generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#pGRR.MSC(y, X)
}
