% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predGRR}
\alias{predGRR}
\title{Prediction}
\usage{
predGRR(res, newX, newZ = NULL)
}
\arguments{
\item{res}{the output of \code{GRR.MSC}, \code{pGRR.MSC}, \code{AM.pGRR}}

\item{newX}{a matrix of penalized explanatory variables for future observation}

\item{newZ}{a matrix of non-penalized explanatory variables, with intercept if need,
for future observation}
}
\value{
predictive values
}
\description{
\code{predGRR} This function provides predictive values for future observation
}
\examples{
#predGRR(res, newX)
}
