% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_pGRR.R
\name{AM.pGRR}
\alias{AM.pGRR}
\title{Additive model via partial generalized ridge regression (v0.3.0)}
\usage{
AM.pGRR(y, X, k = ncol(X), q = k, MSC = "EGCV", alpha = "default", tol = 1e-12)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{k}{the number of explanatory variables}

\item{q}{a vector or an integer which is the number of knots;
if vector, the number of knots is different for each variable}

\item{MSC}{a model selection criterion; see the function \code{GRRMSC}}

\item{alpha}{a value (>=2) expressing penalty strength for \code{MSC}; see the function \code{GRRMSC}}

\item{tol}{tolerance for rank deficient}
}
\value{
a list object with "AM.pGRR" class which has the following elements:
\item{B1, B2}{matrices of explanatory variables transformed by cubic truncated power basis function;
\code{B2} is penalized and \code{B1} is not}

\item{knots}{a matrix of knots}

\item{beta}{estimates for \code{B2}}

\item{gamma}{estimates for \code{B1}}

\item{fitted.values}{fitted values}

\item{R2}{the coefficient of determination}
}
\description{
\code{AM.pGRR} This function estimates additive model via partial generalized ridge regression
}
\examples{
#AM.pGRR(y, X)
}
