% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM_pGRR.R
\name{AM.pGRR}
\alias{AM.pGRR}
\title{Additive model via partial generalized ridge regression}
\usage{
AM.pGRR(y, X, k = ncol(X), q = k)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{k}{the number of explanatory variables}

\item{q}{a vector or an integer which is the number of knots;
if vector, the number of knots is different for each variable}
}
\value{
a list object with "AM.pGRR" class which has the following elements:
\tabular{ll}{
\code{B1}, \code{B2} \tab
matrices of explanatory variables transformed by cubic truncated power basis function;
\code{B2} is penalized and \code{B1} is not \cr \tab \cr
\code{knots} \tab
a matrix of knots \cr \tab \cr
\code{beta} \tab
estimates for \code{B2} \cr \tab \cr
\code{gamma} \tab
estimates for \code{B1} \cr \tab \cr
\code{fitted.values} \tab
fitted values \cr \tab \cr
\code{R2} \tab
the coefficient of determination \cr
}
}
\description{
\code{AM.pGRR} This function estimates additive model via partial generalized ridge regression
}
\examples{
#AM.pGRR(y, X)
}
