% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedModel.R
\name{FM}
\alias{FM}
\title{Generation of simulation data}
\usage{
FM(n, k, rho = 0.99, sig = 1, beta = 1, k0 = floor(k/2))
}
\arguments{
\item{n}{sample size}

\item{k}{the number of explanatory variables}

\item{rho}{correlation between explanatory variables}

\item{sig}{sig^2 is an error variance}

\item{beta}{true regression coefficients (1 or 2)}

\item{k0}{option when beta=1; the number of non-zero coefficients}
}
\value{
simulation data
}
\description{
\code{FM} This function generates simulation data via fixed model
}
\examples{
#FM()
}
