% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRRMSC.R
\name{GRR.MSC}
\alias{GRR.MSC}
\title{Generalized ridge regression optimized based on model selection criterion minimization method}
\usage{
GRR.MSC(
  y,
  X,
  MSC = c("EGCV", "GCp", "GCV", "Cp", "MCp"),
  alpha = log(n),
  n = length(y),
  intercept = TRUE,
  centering = TRUE,
  tol = 1e-12
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{MSC}{a model selection criterion; \code{"EGCV"}, \code{"GCp"}, \code{"GCV"}, \code{"Cp"}, or \code{"MCp"}}

\item{alpha}{a value (>=2) expressing penalty strength for \code{MSC} (only when \code{MSC} is \code{"EGCV"} or \code{"GCp"})}

\item{n}{sample size}

\item{intercept}{if \code{FALSE}, intercept is removed from the model}

\item{centering}{if \code{TRUE}, \code{X} is centralized}

\item{tol}{tolerance for rank deficient}
}
\value{
a list object with "GRR.MSC" class which has the following elements:
\tabular{ll}{
\code{intercept} \tab
the estimate for intercept;
if \code{centering=TRUE}, this is value for centralized explanatory variables \cr \tab \cr
\code{coefOLS} \tab
the ordinary least squares (OLS) estimates \cr \tab \cr
\code{coefGRR} \tab
the generalized ridge regression (GRR) estimates \cr \tab \cr
\code{fitOLS} \tab
the fitted values by OLS \cr \tab \cr
\code{fitGRR} \tab
the fitted values by GRR \cr \tab \cr
\code{theta} \tab
the optimal ridge parameters \cr \tab \cr
\code{R2} \tab
the coefficient of determination \cr \tab \cr
\code{svd} \tab
the output of \code{svd(X)}, where \code{X} is centralized if \code{centering=TRUE} \cr \tab \cr
\code{cand} \tab
candidates of \code{h} which gives the optimal ridge parameters \cr \tab \cr
\code{MSC} \tab
\code{"EGCV"} or \code{"GCp"} used to optimize ridge parameters \cr \tab \cr
\code{alpha} \tab
the value of penalty strength used in \code{MSC} \cr \tab \cr
\code{Xcenter} \tab
if \code{centering=TRUE}, a vector of means for each explanatory variable; if not, NULL \cr
}
}
\description{
\code{GRR.MSC} This function provides generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#GRR.MSC(y, X)
}
