% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRRMSC.R
\name{GRR.MSC}
\alias{GRR.MSC}
\title{Generalized ridge regression optimized based on model selection criterion minimization method (v1.0.0)}
\usage{
GRR.MSC(
  y,
  X,
  MSC = "EGCV",
  alpha = "default",
  n = length(y),
  intercept = TRUE,
  centering = TRUE,
  tol = 1e-12
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{MSC}{a model selection criterion:
\code{"EGCV"} (default), \code{"GCV"}, \code{"GCp"}, \code{"Cp"},\code{"MCp"},
\code{"GIC"}, \code{"AIC"}, \code{"HQC"}, \code{"BIC"}, \code{"AICc"} or \code{"GCVc"}}

\item{alpha}{a value (>=2) expressing penalty strength for \code{"EGCV"}, \code{"GCp"},
\code{"GIC"} or \code{"GCVc"}; default for \code{"EGCV"} and \code{"GIC"} is \code{log(n)};
default for \code{"GCp"} is \code{"optimized"}; default for \code{"GCVc"} is 1}

\item{n}{sample size}

\item{intercept}{if \code{FALSE}, intercept is removed from the model}

\item{centering}{if \code{X} is already centralized, set \code{FALSE}}

\item{tol}{tolerance for rank deficient}
}
\value{
a list object with "GRR.MSC" class which has the following elements:
\item{intercept}{the estimate for intercept;
if \code{centering=TRUE}, this is value for centralized explanatory variables}

\item{coefOLS}{the ordinary least squares (OLS) estimates}

\item{coefGRR}{the generalized ridge regression (GRR) estimates}

\item{fitOLS}{the fitted values by OLS}

\item{fitGRR}{the fitted values by GRR}

\item{theta}{the optimal ridge parameters}

\item{R2}{the coefficient of determination}

\item{svd}{the output of \code{svd(X)}, where \code{X} is centralized if \code{centering=TRUE}}

\item{cand}{candidates of \code{h} which gives the optimal ridge parameters}

\item{MSC}{model selection criterion used to optimize ridge parameters}

\item{alpha}{the value of penalty strength used in \code{MSC}}

\item{Xcenter}{if \code{centering=TRUE}, a vector of means for each explanatory variable; if not, \code{NULL}}
}
\description{
\code{GRR.MSC} This function provides generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#GRR.MSC(y, X)
}
