% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRRMSC.R
\name{GRR.MSC}
\alias{GRR.MSC}
\title{Generalized ridge regression optimized based on model selection criterion minimization method}
\usage{
GRR.MSC(
  y,
  X,
  MSC = c("EGCV", "GCp", "GCV", "Cp", "MCp"),
  alpha = log(n),
  n = length(y),
  centralized = FALSE,
  tol = 1e-12
)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of explanatory variables without intercept}

\item{MSC}{a model selection criterion; EGCV, GCp, GCV, Cp, or MCp}

\item{alpha}{a value (>=2) expressing penalty strength for MSC (only when MSC = "EGCV" or "GCp")}

\item{n}{sample size}

\item{centralized}{TRUE if X is already centralized}

\item{tol}{tolerance for rank deficient}
}
\value{
estimation results
}
\description{
\code{GRR.MSC} This function provides generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#GRR.MSC(y, X)
}
