% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGRRMSC.R
\name{pGRR.MSC}
\alias{pGRR.MSC}
\title{Partial generalized ridge regression optimized based on model selection criterion minimization method (v0.2.0)}
\usage{
pGRR.MSC(y, X, Z, MSC = "EGCV", alpha = "default", n = length(y), tol = 1e-12)
}
\arguments{
\item{y}{a vector of a response variable}

\item{X}{a matrix of penalized explanatory variables without intercept}

\item{Z}{a matrix of non-penalized explanatory variables with intercept if need}

\item{MSC}{a model selection criterion; see the function \code{GRRMSC}}

\item{alpha}{a value (>=2) expressing penalty strength for \code{MSC}; see the function \code{GRRMSC}}

\item{n}{sample size}

\item{tol}{tolerance for rank deficient}
}
\value{
a list object with "pGRR.MSC" class which has the following elements:
\item{beta}{estimates for \code{X}}

\item{gamma}{estimates for \code{Z}}

\item{fitted.values}{fitted values}

\item{R2}{the coefficient of determination}
}
\description{
\code{pGRR.MSC} This function provides partial generalized ridge estimator
with the optimal ridge parameters based on model selection criterion minimization method
}
\examples{
#pGRR.MSC(y, X)
}
